requirements.txt
absl-py==2.1.0
altair==5.3.0
appdirs==1.4.4
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.39
botocore==1.35.39
bs4==0.0.2
cachetools==5.4.0
captcha==0.6.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
cssselect==1.2.0
debugpy==1.8.6
decorator==5.1.1
dnspython==2.7.0
et-xmlfile==1.1.0
executing==2.1.0
extra-streamlit-components==0.1.71
fake-useragent==1.5.1
flatbuffers==24.3.25
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
googleapis-common-protos==1.63.2
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
idna==3.7
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.4.1
libclang==18.1.1
lxml==5.3.0
lxml_html_clean==0.3.1
lz4==4.3.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.0
namex==0.0.8
nest-asyncio==1.6.0
numpy==1.26.4
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parse==1.20.2
parso==0.8.4
perception==0.6.8
pillow==10.4.0
pinecone==5.4.0
pinecone-client==4.1.2
pinecone-plugin-inference==2.0.1
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
progiter==2.0.0
prompt_toolkit==3.0.48
protobuf==4.25.3
protoc-gen-openapiv2==0.0.1
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
pydeck==0.9.1
pyee==11.1.1
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.10.1
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic-bin==0.4.14
pytz==2024.1
PyWavelets==1.6.0
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-html==0.10.0
rich==13.7.1
rpds-py==0.19.0
s3transfer==0.10.3
scipy==1.14.0
selenium==4.23.0
setuptools==70.3.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.1
streamlit==1.39.0
streamlit-authenticator==0.4.1
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-hub==0.16.1
tensorflow-intel==2.17.0
termcolor==2.4.0
tf_keras==2.17.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.20
validators==0.33.0
w3lib==2.2.1
watchdog==4.0.1
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
wsproto==1.2.0
zipp==3.20.2
